+++ tests/mocks/autoloader.php
@@ -44,6 +44,7 @@ function autoload($class)
 		'Migration',
 		'Pagination',
 		'Parser',
+		'Password_generator',
 		'Profiler',
 		'Table',
 		'Trackback',